# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Okten test docs",
    "version": "1.0"
  },
  "paths": {
    "/users/{userId}": {
      "get": {
        "description": "Get user by ID",
        "tags": ["User"],
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "type": "string",
            "required": true
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/SingleUser"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      },
      "put": {
        "description": "Update user",
        "tags": ["User"],
        "operationId": "updateUserById",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "type": "string",
            "required": true
          },
          {
            "name": "name",
            "required": false,
            "type": "string",
            "in": "formData"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "User was updated",
            "schema": {
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/SingleUser"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "description": "Get all users",
        "tags": ["User"],
        "operationId": "getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SingleUser"
              }
            }
          }
        }
      }
    }
  },

  "definitions": {
    "ApiError": {
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 400
        },
        "message": {
          "type": "string",
          "example": "Bad request"
        }
      }
    },
    "SingleUser": {
      "properties": {
        "name": {
          "type": "string",
          "example": "Denys"
        },
        "age": {
          "type": "integer",
          "example": 21
        },
        "email": {
          "type": "string",
          "example": "example@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "123asd"
        }
      }
    }
  }
}